refactor: move errors and response utils
