+++ src/utils/errors/errorhandler.js
@@ -1,17 +1,17 @@
-export const errorHandler = (err, req, res) => {
+export default (err, req, res) => {
   if (err.statusCode) {
     // if error has user-defined statusCode then it's a custom error
     res.status(err.statusCode).json({
       status: "error",
-      error: err.message,
-      data: err.data,
+      message: err.message,
+      errors: err.errors,
     });
     // if error has system-generated status
   } else if (err.status) {
     res.status(err.status).json({
       status: "error",
-      error: err.message,
-      data: [], // no data to return
+      message: err.message,
+      errors: [], // no data to return
     });
     // if this is an unknown/uncaught error
   } else {
@@ -22,8 +22,8 @@ export const errorHandler = (err, req, res) => {
     }
     res.status(500).json({
       status: "error",
-      error: `Internal server error${serverError}`,
-      data: [], // no data to return
+      message: `Internal server error${serverError}`,
+      errors: [], // no data to return
     });
     //we want to see the actual error
     console.log(err.message);
